+++ src/DescriptorSystems.jl
@@ -14,13 +14,13 @@ import Polynomials: AbstractRationalFunction, AbstractPolynomial, poles, isconst
 
 export DescriptorStateSpace, AbstractDescriptorStateSpace, dss, dssdata, rdss, rss, iszero, order
 export RationalTransferFunction, rtf
-export gminreal, gir, gir_lrtran, gbalmr, gsvselect, gss2ss, dss2ss
+export gminreal, gir, gir_lrtran, gbalmr, dsxvarsel, gss2ss, dss2ss
 export gbilin, c2d, dss2rm, dss2pm, timeresp
 export confmap, simplify, normalize, poles, gain, zpk, rtfbilin, numpoly, denpoly, isconstant, sampling_time
 export blockdiag, eye, rcond
 export gdual, ctranspose, inv, ldiv, rdiv, ginv
 export append, series, parallel, horzcat, vertcat
-export order, evalfr, dcgain, opnorm, freqresp, chess
+export order, evalfr, dcgain, opnorm, freqresp, chess, dssubset, dszeros, dscopy, dssubsel, dsdiag
 export gpole, gzero, gpoleinfo, gzeroinfo, gnrank, isregular, isproper, isstable, 
        glinfnorm, ghinfnorm, gl2norm, gh2norm, ghanorm, gnugap
 export gsdec, grnull, glnull, grange, gcrange, grsol, glsol, grmcover1, grmcover2, glmcover1, glmcover2